refactor(core): extract user search query builder
