+++ packages/core/src/modules/users/services/users.service.ts
@@ -10,6 +10,7 @@ import { CreateUserModel } from "../models/create-user.model";
 import { PermissionsService } from "./permissions.service";
 import { RolesService } from "./roles.service";
 import argon2 from "argon2";
+import { buildUserSearchQuery } from "../helpers/build-user-search-query";
 
 @Injectable()
 export class UserService {
@@ -54,59 +55,7 @@ export class UserService {
    */
   async countUsers(filters?: Record<string, any>): Promise<number> {
     try {
-      const { search, ...otherFilters } = filters || {};
-      const query: any = { ...otherFilters, deletedAt: null };
-
-      if (search && typeof search === 'string' && search.trim()) {
-        const searchTerm = search.trim();
-
-        const searchWords = searchTerm.split(/\s+/).filter(word => word.length > 0);
-
-        if (searchWords.length === 1) {
-          const singleWordRegex = new RegExp(searchWords[0], 'i');
-          query.$or = [
-            { firstName: { $regex: singleWordRegex } },
-            { lastName: { $regex: singleWordRegex } },
-            { email: { $regex: singleWordRegex } }
-          ];
-        } else if (searchWords.length >= 2) {
-          const [firstWord, secondWord, ...otherWords] = searchWords;
-          const firstWordRegex = new RegExp(firstWord, 'i');
-          const secondWordRegex = new RegExp(secondWord, 'i');
-          const allWordsRegex = searchWords.map(word => new RegExp(word, 'i'));
-
-          query.$or = [
-            { firstName: { $regex: new RegExp(searchTerm, 'i') } },
-            { lastName: { $regex: new RegExp(searchTerm, 'i') } },
-            { email: { $regex: new RegExp(searchTerm, 'i') } },
-            {
-              $and: [
-                { firstName: { $regex: firstWordRegex } },
-                { lastName: { $regex: secondWordRegex } }
-              ]
-            },
-            {
-              $and: [
-                { firstName: { $regex: secondWordRegex } },
-                { lastName: { $regex: firstWordRegex } }
-              ]
-            },
-
-            ...(otherWords.length > 0 ? [{
-              $and: allWordsRegex.map(regex => ({
-                $or: [
-                  { firstName: { $regex: regex } },
-                  { lastName: { $regex: regex } },
-                  { email: { $regex: regex } }
-                ]
-              }))
-            }] : []),
-
-            { email: { $regex: new RegExp(searchWords.join('|'), 'i') } }
-          ];
-        }
-      }
-
+      const query = buildUserSearchQuery(filters);
       return await this.userModel.countDocuments(query).exec();
     } catch (error) {
       const errorMessage =
@@ -130,66 +79,7 @@ export class UserService {
     filters?: Record<string, any>,
   ): Promise<UserResponseModel[]> {
     try {
-      // Estrai il parametro search dai filtri
-      const { search, ...otherFilters } = filters || {};
-
-      // Costruisci la query base
-      const query: any = { ...otherFilters, deletedAt: null };
-
-      // Se c'è un parametro search, aggiungi le condizioni di ricerca
-      if (search && typeof search === 'string' && search.trim()) {
-        const searchTerm = search.trim();
-
-        // Dividi il termine di ricerca in parole
-        const searchWords = searchTerm.split(/\s+/).filter(word => word.length > 0);
-
-        if (searchWords.length === 1) {
-          // Ricerca con una sola parola
-          const singleWordRegex = new RegExp(searchWords[0], 'i');
-          query.$or = [
-            { firstName: { $regex: singleWordRegex } },
-            { lastName: { $regex: singleWordRegex } },
-            { email: { $regex: singleWordRegex } }
-          ];
-        } else if (searchWords.length >= 2) {
-          // Ricerca con più parole - gestisce sia "Mario Rossi" che "Rossi Mario"
-          const [firstWord, secondWord, ...otherWords] = searchWords;
-
-          const firstWordRegex = new RegExp(firstWord, 'i');
-          const secondWordRegex = new RegExp(secondWord, 'i');
-          const allWordsRegex = searchWords.map(word => new RegExp(word, 'i'));
-
-          query.$or = [
-            { firstName: { $regex: new RegExp(searchTerm, 'i') } },
-            { lastName: { $regex: new RegExp(searchTerm, 'i') } },
-            { email: { $regex: new RegExp(searchTerm, 'i') } },
-            {
-              $and: [
-                { firstName: { $regex: firstWordRegex } },
-                { lastName: { $regex: secondWordRegex } }
-              ]
-            },
-            {
-              $and: [
-                { firstName: { $regex: secondWordRegex } },
-                { lastName: { $regex: firstWordRegex } }
-              ]
-            },
-
-            ...(otherWords.length > 0 ? [{
-              $and: allWordsRegex.map(regex => ({
-                $or: [
-                  { firstName: { $regex: regex } },
-                  { lastName: { $regex: regex } },
-                  { email: { $regex: regex } }
-                ]
-              }))
-            }] : []),
-
-            { email: { $regex: new RegExp(searchWords.join('|'), 'i') } }
-          ];
-        }
-      }
+      const query = buildUserSearchQuery(filters);
 
       const users = await this.userModel
         .find(query)
